restore stub and fix phpstan
