+++ src/Handlers/MainCommand.php
@@ -7,8 +7,6 @@ namespace Oscabrera\ModelRepository\Handlers;
 use Illuminate\Console\Command;
 use Illuminate\Console\View\Components\Factory;
 use Oscabrera\ModelRepository\Classes\Options;
-use Oscabrera\ModelRepository\Exception\Command\CreateStructureException;
-use Oscabrera\ModelRepository\Exception\Command\StubException;
 use Oscabrera\ModelRepository\Handlers\Makers\MakeController;
 use Oscabrera\ModelRepository\Handlers\Makers\MakeInterfaceRepository;
 use Oscabrera\ModelRepository\Handlers\Makers\MakeInterfaceServices;
@@ -16,9 +14,12 @@ use Oscabrera\ModelRepository\Handlers\Makers\MakeModel;
 use Oscabrera\ModelRepository\Handlers\Makers\MakeRepository;
 use Oscabrera\ModelRepository\Handlers\Makers\MakeRequest;
 use Oscabrera\ModelRepository\Handlers\Makers\MakeService;
+use Oscabrera\ModelRepository\Trait\MainCommandTrait;
 
 class MainCommand
 {
+    use MainCommandTrait;
+
     protected Factory $output;
 
     protected Command $command;
@@ -26,14 +27,6 @@ class MainCommand
     protected Options $options;
 
     protected string $name;
-
-    /**
-     * @var array<int, class-string>
-     */
-    private array $hasBinding = [
-        MakeRepository::class,
-        MakeService::class,
-    ];
 
     /**
      * Constructor for the class.
@@ -177,28 +170,6 @@ class MainCommand
     {
         if ($this->options->hasRequest()) {
             $this->executeMaker($this->makeRequest);
-        }
-    }
-
-    /**
-     * Executes the Maker.
-     *
-     * This method executes a Maker by calling its `make` method with the given options.
-     * If the `bind` parameter is set to `true`, it also calls the Maker's `binding` method.
-     */
-    private function executeMaker(
-        MakeRepository|MakeInterfaceRepository|MakeService|MakeInterfaceServices|MakeController|MakeRequest $maker,
-    ): void {
-        try {
-            $result = $maker->make($this->name, $this->options);
-            if (in_array(get_class($maker), $this->hasBinding, true)) {
-                $maker->binding($this->name);
-            }
-            $this->infoCommand($result);
-        } catch (StubException|CreateStructureException $exception) {
-            $info = $exception->getInput();
-            /** @var array{type: string, path: string} $info */
-            $this->errorCommand($exception->getMessage(), $info);
         }
     }
 }
