+++ stubs/Service.stub
@@ -19,6 +19,8 @@ class DummyClass implements DummyInterface
 {
     /**
      * Class Constructor.
+     *
+     * @param DummyRepository $repository The model object to be injected.
      */
     public function __construct(
         protected DummyRepository $repository
@@ -27,6 +29,11 @@ class DummyClass implements DummyInterface
 
     /**
      * Create a new DummyModel resource.
+     *
+     * @param DummyModelRequest $request The request object containing the data for creating a DummyModel.
+     *
+     * @return EntityResourceTransformer Returns the newly created UserResource if successful,
+     * or null if the User creation failed.
      */
     public function create(
         DummyModelRequest $request
@@ -38,6 +45,10 @@ class DummyClass implements DummyInterface
 
     /**
      * Read a DummyModel resource by ID.
+     *
+     * @param string $id The ID of the DummyModel resource to fetch.
+     *
+     * @return EntityResourceTransformer The DummyModel resource, or null if not found.
      */
     public function read(string $id): EntityResourceTransformer
     {
@@ -46,6 +57,11 @@ class DummyClass implements DummyInterface
 
     /**
      * Updates a DummyModel resource.
+     *
+     * @param string $id The ID of the DummyModel resource to update.
+     * @param DummyModelRequest $request The request data for updating the DummyModel.
+     *
+     * @return bool
      */
     public function update(string $id, DummyModelRequest $request): bool
     {
@@ -57,6 +73,10 @@ class DummyClass implements DummyInterface
 
     /**
      * Deletes a DummyModel object based on the provided ID.
+     *
+     * @param string $id The ID of the DummyModel object to delete.
+     *
+     * @return bool True if the deletion was successful, false otherwise.
      */
     public function delete(string $id): bool
     {
@@ -65,6 +85,8 @@ class DummyClass implements DummyInterface
 
     /**
      * List all DummyModel resources.
+     *
+     * @return JsonApiPaginationTransformer
      */
     public function list(): JsonApiPaginationTransformer
     {
